move schema.ts and stream-object.ts from v5 -> base
