+++ packages/core/src/stream/base/schema.ts
